+++ packages/core/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./hash";
 export * from "./lifecycle";
+export * from "./module-api";
 export * from "./schema-validation";
 export * from "./state";
 export * from "./status";